chat-bot/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.19.0",
    "fastapi>=0.109.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.26.0",
    "openai>=1.12.0",
    "pydantic>=2.5.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.7.1",
    "uvicorn>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]
